+++ src/Procon.Database.Serialization/Serializers/NoSql/SerializerMongoDb.cs
@@ -98,8 +98,8 @@ namespace Procon.Database.Serialization.Serializers.NoSql {
                 if (numeric.Long.HasValue == true) {
                     parsed = numeric.Long.Value.ToString(CultureInfo.InvariantCulture);
                 }
-                else if (numeric.Float.HasValue == true) {
-                    parsed = numeric.Float.Value.ToString(CultureInfo.InvariantCulture);
+                else if (numeric.Double.HasValue == true) {
+                    parsed = numeric.Double.Value.ToString(CultureInfo.InvariantCulture);
                 }
             }
             else if (@string != null) {
